requirements.txt
google-auth-oauthlib==0.2.0
pytz==2018.3
bigquery-python
